+++ base/docs/basedocs.jl
@@ -3253,7 +3253,9 @@ Array{T,N}(::Missing, dims)
 Singleton type used in array initialization, indicating the array-constructor-caller
 would like an uninitialized array.
 
-See also [`undef`](@ref), an alias for `UndefInitializer()`.
+See the section in the manual on uninitialized memory.
+
+See also: [`undef`](@ref), an alias for `UndefInitializer()`.
 
 # Examples
 ```julia-repl
@@ -3273,7 +3275,9 @@ Alias for `UndefInitializer()`, which constructs an instance of the singleton ty
 [`UndefInitializer`](@ref), used in array initialization to indicate the
 array-constructor-caller would like an uninitialized array.
 
-See also [`missing`](@ref), [`similar`](@ref).
+See the section in the manual on uninitialized memory.
+
+See also: [`missing`](@ref), [`similar`](@ref).
 
 # Examples
 ```julia-repl